Solved cyclic dependencies, added Provider Interface
