+++ ssolibs/samllib/src/test/resources/Saml/valid/Saml1.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<saml
+xmlns="http://www.api.sso.nds.rub.org">
+</saml>
