+++ ssolibs/samllib/src/test/resources/Saml/valid/Saml6.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<saml
+xmlns="http://www.api.sso.nds.rub.org">
+    <samlResponse>
+        PHNhbWxwOlJlc3BvbnNlIHhtbG5zOnNhbWxwPSJ1cm46b2FzaXM6bmFtZXM6dGM6U0FNTDoyLjA6cHJvdG9jb2wiIHhtbG5zOnNhbWw9InVybjpvYXNpczpuYW1lczp0YzpTQU1MOjIuMDphc3NlcnRpb24iIElEPSJfOGU4ZGM1ZjY5YTk4Y2M0YzFmZjM0MjdlNWNlMzQ2MDZmZDY3MmY5MWU2IiBWZXJzaW9uPSIyLjAiIElzc3VlSW5zdGFudD0iMjAxNC0wNy0xN1QwMTowMTo0OFoiIERlc3RpbmF0aW9uPSJodHRwOi8vc3AuZXhhbXBsZS5jb20vZGVtbzEvaW5kZXgucGhwP2FjcyIgSW5SZXNwb25zZVRvPSJPTkVMT0dJTl80ZmVlM2IwNDYzOTVjNGU3NTEwMTFlOTdmODkwMGI1MjczZDU2Njg1Ij4NCiAgPHNhbWw6SXNzdWVyPmh0dHA6Ly9pZHAuZXhhbXBsZS5jb20vbWV0YWRhdGEucGhwPC9zYW1sOklzc3Vlcj4NCiAgPHNhbWxwOlN0YXR1cz4NCiAgICA8c2FtbHA6U3RhdHVzQ29kZSBWYWx1ZT0idXJuOm9hc2lzOm5hbWVzOnRjOlNBTUw6Mi4wOnN0YXR1czpTdWNjZXNzIi8+DQogIDwvc2FtbHA6U3RhdHVzPg0KICA8c2FtbDpFbmNyeXB0ZWRBc3NlcnRpb24+DQogICAgPHhlbmM6RW5jcnlwdGVkRGF0YSB4bWxuczp4ZW5jPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxLzA0L3htbGVuYyMiIHhtbG5zOmRzaWc9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvMDkveG1sZHNpZyMiIFR5cGU9Imh0dHA6Ly93d3cudzMub3JnLzIwMDEvMDQveG1sZW5jI0VsZW1lbnQiPjx4ZW5jOkVuY3J5cHRpb25NZXRob2QgQWxnb3JpdGhtPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxLzA0L3htbGVuYyNhZXMxMjgtY2JjIi8+PGRzaWc6S2V5SW5mbyB4bWxuczpkc2lnPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwLzA5L3htbGRzaWcjIj48eGVuYzpFbmNyeXB0ZWRLZXk+PHhlbmM6RW5jcnlwdGlvbk1ldGhvZCBBbGdvcml0aG09Imh0dHA6Ly93d3cudzMub3JnLzIwMDEvMDQveG1sZW5jI3JzYS0xXzUiLz48eGVuYzpDaXBoZXJEYXRhPjx4ZW5jOkNpcGhlclZhbHVlPmZMM3cwSS9UQ1Q2Z09VaVd6V3JkYjFPQnFjYTB2WnpxQ2JnMTIzMERqWDdRazBjb0UwV1JuK2ZjNlhKRWR5Mk1vL0M5NGY2dkRGdWFOWWdNN1RRalBwUzVyZUJUQy9EWmFRWHE4ekx2UnUzUEdiK09jcXRiRGdkNzRiSUR4Rll5ZVA0YzJJSG1TeElvWjZwWXNEZ0NwYXpFQ0h6MzBBb29aQkZKYTU0WTl4cz08L3hlbmM6Q2lwaGVyVmFsdWU+PC94ZW5jOkNpcGhlckRhdGE+PC94ZW5jOkVuY3J5cHRlZEtleT48L2RzaWc6S2V5SW5mbz4NCiAgIDx4ZW5jOkNpcGhlckRhdGE+DQogICAgICA8eGVuYzpDaXBoZXJWYWx1ZT5LekExOG9BNkgrd0FrNjg0WHUwb0h2dG1WNmJndlNoKzBvRzNVZkhSVHlpeGFNbGNoMjNLVGlqRG9KYThaS1RCTkxkbGxWZE1MYWV6TjY3QWFESVRCNE1CdzA0bk5YZ21kZ0VLNThCaDBHVjRjUjJ6anZRa0FzRE0yMXJKM3RaNmNobU43Ykh2U1pJM1kvS2x6Q1IxQzB4Sml2YjdRQlF5L283aDR5cDNwdVI5M1pHR242S3ZKUzdPY1UzU1d2bWNGMTVpa1dWZzk5Ryt6ZGIwQ3RTQlFLQWdpRWlsYjdxdFBJN2swU0pLTnFvRnBPa3R5b29MUTVvT0lCVkQ5M3RkUE43VmNoNUFJakNUQXNiTkhuRjE4eWYyOXFlazNiKzNSOU9rc29yMEN6UldnVTVSY1p0cFpiZEd5dEI0NkRGYzZGWFBLSXcxWkdjbVAzVGVBTTZPV2ZibWZHLzJCMEhNYzhudmtYUjhleXEwVE1idy9EMG1qNXJtNTBma3F5T1JQcm51eDltMDJibDRUQW1FS3lUTzhYMER5bDJZRTFndWJpMlVpNDA5QXRQeHhOaG1MekN3SUVOSWJtSTZuSU84aWpiaVM3MGFiemlSV1NrUlZwVUFXVWYwYmJ6SU9lbnQ0MWoySm8wUWpkRjFHdmtmYWlqSmFCNlB2cDA5aURXdVVZWjUzOFBpd3gyQzJXWk1QZ0QvWXVWWHc5YVBtWTExNzU5K2tZWHRrWFk5MUFvTzZ6TjB5YUVzVDhIOCtzZXNVdE1Sd05EWGdPY0oxRG1jT2FTbExjQkhRMUszNjQ5QTE0NGJmS1hiUzhqK2hlUTdOVENhRG1TeTFsbnkwWm1jYVNxMXR4bUR4NnkrZzhVSVhoWnR0Y3ptZWExSVVLN1h1bmNoYTRXam1MbTgyK3pkTklCM3lpSUFjaHlvQktGRHlsUzJqRjdiVElDcytoUnRwMXZDS0JtYkJQcjNRYWswMUNFOXVQMm4wMFlER3Zrc05ZMmUxclRtSnAwMTlnQkhsSkZ4eTdIMTBHU0dmdXJQVlFwR3ZUY2FHSTA1NlBlU1NXZGo5clJxa0NjWFI0QmZDbDZGQ29Ha3lFN2M2TjlwcjA4N0dheDVXY0V4V3FYTEtsa1lLWmNENy8xVXcwWWhtQTJJbVRySTBDZVNuSTFhbnZHQ0Q2WWUxekMvbnY0RzZ6d1I1KytPUzU5VWpSUm12a2NCUmVkeDNOampXaFM4SU1KZURVajZXU2EyTFl5SlJBZm1xQmluVW1sdTljcnZOYldyck1VeDYya0dmSzFHaEhkVnpxdVhUVXhTbGloMWtINUNXS0tIclRSU3grVmFGaGUyV3FsWUorUzh5WWg2L24zMlRVbk1CKzFNdmduZmJ0TlVUdE93MUYwTXRJSGRrZkllbzAvSHhmRThiaFpqdHJxZ2hEcmxwZkJ5d1YyQmNxdzZ0VFc5REFMSHh3eGJuQ3FnUjNFSHlkMTJSUFhydE02Ni9RMWVjaEpBbnFDNlJvQ1VkWm5MeS9Uc1NtR200RWpSK2Yvakhwdnd0MEdiSm5VU1paMGxLRjBpZzRwK1RLZUZITzY2TEJEVVRrRWZVMTQyVG5BRTBMaG51bXFTTTdjZFY5WHRtQnN5Qmt2UWQ4UGd5aWk5b3VEdFgxM1ZON2Vlb2N5OWcrMDlSN2ZtUUtmdEV1VEZlVEM4WlBIeEN3QlhHbmUwanVKem5jd0V4VUZGMW5EV21haTlQVVdMOFExL2p5Mmh4TUZjNjNCQmkxSkdqREpUY1VpRkF3dWM5SmgxTE1CTXhsbnRVblZSNlRib2ZVdTdnSnhhS3g4TEdNblVVYmNTdkZ0QW5RWktJUk9jdE9vK1FrTmNhZkVGWFY0aG5IaFpzRU5ldTNDd28zSEI5S0t1YzVwVUpBRmJqcDJwUE1xbnU0cnM5eU8zc295Q1ZmTXd3OVNhdnp0NHlZQUxNRVJUQnkyc1pJdXJueEtLNmIwVDNkWXA4UUFNc1QzRjlEb1luUG5hTFUzaG5Ib1N5Wjgrb0ZNLzVXVXdma21JNUsvV1Ivd1FEZGJ1UHB0MjBBNjQyZ2Q3R0NIeWt0ZnNNWnBGS3VTMTRySTRnaUdMVmhNRjMzMzdHRmJBT3d6cG9IVms0K1I2RHQ2eTNTY0xBNC9Vc2kyTHVjNFJ4eUtkc21vMEZpYzFnRHR5am02K1AzbjRTNzIwZHBVNUp3MGxlZlhYWkxrWFRVc1k1WG9MVTZLdzBQV3FZWHFGdERpczhjc2NNWXFsVzVNTEU1bnNtcStiQVdIaktQcjRYT01Edkc3Zk1rekdRVTdQd3pleFVyUFlEU2JnRmFnMnBtTHp6cnhyL0JKRlBSSVQwa0xPK3JWWkpaUHpPSno1a1ZCVXljL1licmUydWZWd2hBL2hNODZYKzh5U3RnbXpIZkMveGU4Ui84ajlMdDA0dmNsbXpTZ25XdmdCS1libWRYSlR5NGViVWlHL3R1N0JIYlByTmRIV21RWkJxeGpva3FXZGlhVHlLMWh6eWgzUzRRQStrZytpSGJ2dldTSmRhWUNtSWFtdFV0Rk11cmVacjJGSExiMDQvMHVGVnI2MHBEanBFVlpUZWtvZEt2ZFhqTVF1YURIeUFtN0xleFFYV001QnpodzIzQmYwVWJOa2dRNzFiN2ZTbndOZnlKVUhXSFZDcTNocUYrNDVWdlJTWjNXYnRJRVJsWTFmdTFtaDNBdHBkUzJoeFB5VVZKQUFPbGtsUHpIZkZ5cUFEbklYM2MwMU9VQTdXY201M1Myd2F1cThBNFhLL0FIM3FRTW9xd2JmbXdwTXdwc21jeWRCM3pWdTFEVERXT0FmVzBJS0hwZmdoMm01Znd4OGdXVlhmUWdheC9mbW8xWDFKWkk3a21VdTM3cGdJQUJGSHlLSU5vc2JmYjY1M3Q4bzM5SnY4NWhnLzA3djVyUG5mRkRLSktSYnRCZmJEQW5adE5EbzdPckVhZUpQMFlWSHVxN0E2RmlVbVBLSXNnZWdFT0Rkd3Y2eWVqWFY4bGtkelVxUm03NTBNRnQ2NDUxS01LUmNVazUwYWhwZzRnUmo0SzI1TEx1M0FWeXN5cFBHVVA0OUN6Q2N3WWFrZzhvWnlaWGVTQktQNFRYUElTTUpVbDR4TENWd3F6Z1BZcHdZdTNvU3k0TG5KYWZlWmNrU2lzSkY4Qjdobm1TRjZYWUVQR3FBQ0tMMGlrSlliU2czN2ljMnBBQmxpa0lScGlSenJZVDJabWVJUzJnSlVROURvOUl0Yy9NTVNYZlhOTS9oUGVNRmVuZlFYdW1yd2tlOWxNZ05DY2ZZbXBrTC9OT2VFZ2JoOC8rMXJVazNqODkxZWMrSzdwYVV3aTFPS1dMSVBNZmZNNFVNRkUyY01LQzZmWFB6aG93NGRtTHd6a0tvdkh5MkRzRG1xOEZtc0ZXNUxaVjR1a014RVYxeW4xZExjdDJYTHR4L0ZuY0Nxd0hCdW5Lbk9YMEI3bzFnTEV5SmdESDYyc2p4VVVjN1BqZ2dpeXNiSjkzV0tia1hZamRqRm9pK2k0bWNaK1lGZEQxTEwraWxOUldyRStOZmJjZzRsRU91SElLTWJhVWxrVnhYSWNmZkwwaFFNZnBWdGU0M2tHRHFGQUExTVNOUjVTUjIxSUV4R2VNZ1VONDFJMWZjUUw3Zlp6QmJWamZiM0VEUUJCYXVaQ2FuNVB3Y3ZIWFExcFM1Q1haOXJXbGFuQVI4ajdCR0NCaENFR21MSGJHaThxNXFadnFxZ09ScG9VUjV0aG5rT3J0YnZLMll6RmM5SWtCQjErNlhDTmhTNWlpS2VvUHJXMXVNQXpCMVpmNFFUQW1HVmRuTVJTbWhvU3Boc25PSlplK2IxWFZhZW5jPTwveGVuYzpDaXBoZXJWYWx1ZT4NCiAgIDwveGVuYzpDaXBoZXJEYXRhPg0KPC94ZW5jOkVuY3J5cHRlZERhdGE+DQogIDwvc2FtbDpFbmNyeXB0ZWRBc3NlcnRpb24+DQo8L3NhbWxwOlJlc3BvbnNlPg==
+    </samlResponse>
+</saml>
