+++ ssolibs/sso_api/pom.xml
@@ -15,11 +15,6 @@
             <artifactId>utils</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
-        <dependency>
-            <groupId>org.rub.nds.sso</groupId>
-            <artifactId>samllib</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
     </dependencies>
     
     <properties>
